seperated all chainable and notchainable lodash mixins
